# .futurxlab/api-specification.yaml
openapi: 3.0.3
info:
  title: InKnowing API
  description: |
    InKnowing is an AI-powered book dialogue platform that enables users to have conversations with books and book characters.

    ## Business Logic Conservation
    Based on the principle of business logic conservation, this API documentation maintains complete mapping between:
    - User Journey → API Endpoints
    - State Transitions → API Operations
    - Sequence Interactions → API Calls
    - Business Rules → API Responses

    ## Key Features
    - Question-driven and book-driven discovery
    - AI-powered dialogues with books and characters
    - User book upload with vectorization
    - Multi-tier membership system
    - Admin management dashboard

    ## Authentication
    The API uses JWT Bearer tokens for authentication. Tokens are obtained through the login endpoints and must be included in the Authorization header for protected endpoints.

    ## Rate Limiting
    - Free users: 20 requests per day
    - Basic members: 200 requests per month
    - Premium members: 500 requests per month
    - Super members: 1000 requests per month
  version: 1.0.0
  contact:
    name: InKnowing Support
    email: support@inknowing.ai
    url: https://inknowing.ai/support
  license:
    name: Proprietary
    url: https://inknowing.ai/terms
servers:
  - url: https://api.inknowing.ai/v1
    description: Production server
  - url: https://staging-api.inknowing.ai/v1
    description: Staging server
...
paths:
...
  /search:
    get:
      tags:
        - Search
      summary: Search books by question or query
      description: |
        Search for books that can answer a question or match a query
        Business Logic: Question → Book Discovery
      parameters:
        - name: q
          in: query
          required: true
          description: Search query or question
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: type
          in: query
          description: Search type
          schema:
            type: string
            enum: [question, title, author]
            default: question
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
  /search/books:
    get:
      tags:
        - Search
      summary: Search books by title
      description: Direct book title search
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
        - name: exact
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Book search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookList'

  # ==================== Books ====================
...
  /uploads/check:
    post:
      tags:
        - Uploads
      summary: Check if book exists
      description: Check if a book already exists before upload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                author:
                  type: string
      responses:
        '200':
          description: Check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
                  book_id:
                    type: string
                    nullable: true
                  ai_known:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
  /uploads:
    post:
      tags:
        - Uploads
      summary: Upload book
      description: |
        Upload a book file for processing
        Business Logic: User Upload → AI Processing → Vectorization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - title
                - author
              properties:
                file:
                  type: string
                  format: binary
                  description: Book file (TXT or PDF, max 10MB)
                title:
                  type: string
                author:
                  type: string
                category:
                  type: string
                  enum: [business, psychology, fiction, science, history, philosophy]
                description:
                  type: string
                  maxLength: 200
      responses:
        '202':
          description: Upload accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Upload'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: File too large
  /uploads/{uploadId}:
    get:
      tags:
        - Uploads
      summary: Get upload status
      description: Get processing status of uploaded book
      security:
        - bearerAuth: []
      parameters:
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Upload status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Upload'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /uploads/my:
    get:
      tags:
        - Uploads
      summary: Get user uploads
      description: Get list of books uploaded by current user
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed, all]
            default: all
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: User uploads
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploads:
                    type: array
                    items:
                      $ref: '#/components/schemas/Upload'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Admin ====================
...
components:
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        cover:
          type: string
          nullable: true
        category:
          type: string
        description:
          type: string
        dialogue_count:
          type: integer
        rating:
          type: number
          minimum: 0
          maximum: 5
        created_at:
          type: string
          format: date-time
...
    BookList:
      type: object
      properties:
        books:
          type: array
          items:
            $ref: '#/components/schemas/Book'
        pagination:
          $ref: '#/components/schemas/Pagination'
    SearchResults:
      type: object
      properties:
        query:
          type: string
        type:
          type: string
          enum: [question, title, author]
        results:
          type: array
          items:
            type: object
            properties:
              book:
                $ref: '#/components/schemas/Book'
              relevance_score:
                type: number
                minimum: 0
                maximum: 100
              matched_chapters:
                type: array
                items:
                  type: object
                  properties:
                    chapter_number:
                      type: integer
                    chapter_title:
                      type: string
                    preview:
                      type: string
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ==================== Character Schemas ====================
...
    Upload:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        book_id:
          type: string
          nullable: true
          description: Book ID once processing is complete
        filename:
          type: string
        file_size:
          type: integer
        file_type:
          type: string
          enum: [txt, pdf]
        title:
          type: string
        author:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        processing_steps:
          type: array
          items:
            type: object
            properties:
              step:
                type: string
                enum: [
                  ai_detection,
                  text_preprocessing,
                  chapter_extraction,
                  character_extraction,
                  vectorization,
                  indexing,
                  model_generation
                ]
              status:
                type: string
                enum: [pending, processing, completed, failed]
              progress:
                type: integer
                minimum: 0
                maximum: 100
              message:
                type: string
        ai_known:
          type: boolean
          nullable: true
        vector_count:
          type: integer
          nullable: true
        extracted_characters:
          type: array
          items:
            type: string
        points_earned:
          type: integer
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true

    # ==================== AI Model Schemas ====================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication

    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication

# API Response Standards
...